feat/refactor-components: extract dashboard folder from the componnents folder to the app folder
