+++ app/dashboard/search-bar/index.tsx
